requirements.txt
openai
python-dotenv
pandas
numpy
scikit-learn
nltk
